include rehype-slug behaviour, move to individual file
