+++ package.json
@@ -69,7 +69,6 @@
 		"rehype-autolink-headings": "^7.1.0",
 		"rehype-external-links": "^3.0.0",
 		"rehype-mermaid": "^2.1.0",
-		"rehype-slug": "^6.0.0",
 		"rehype-title-figure": "^0.1.2",
 		"sharp": "^0.33.5",
 		"solarflare-theme": "^0.0.2",
